Fix usage of optional rn variable which was not being passed.
Optional argument rn was not being checked for presence when values were being assigned to it.
Additionally a call did not pass in the argument, which caused a failure when the value was assigned to rn

Solution : Pass in argument to all calls and remove optional capability.
Variable rn is a useful diagnostic in both cases.