Added logic to guard against incorrect use of OH tracer array.
Previously if the OH tracer was not in the field table, an array (oh_conc) was being used even though it was badly
indexed from the caller routine (as nOH < 0 ). The value calculated using this array is only used if
do_dynamic_bc or do_dynamic_om are true, so added logic to only use oh_conc if one of these was true.

Also added logic to initialization to trap that the OH tracer is present if either is true.