Add new module_flux_is_initialized to atmos_co2.F90
Corrects an error when the namelist would not be read if the
atmos_co2_flux_init subroutine was called before the atmos_co2_init
subroutine in atmos_co2.F90.

The fix is to create a new logical variable module_flux_is_initialized
to be used by atmos_co2_flux_init.

Fixes #4